JS-3288: default type in space settings
